Cleaning up fake data service
