+++ app/views/api/v3/analytics/user_analytics/_period_report.html.erb
@@ -3,7 +3,7 @@
     <a
       class="d-block pl-16px mb-24px"
       href="#"
-      onclick="goToPage('<%= pageId %>', 'home-page'); return false;"
+      onclick="goToPage('<%= page_id %>', 'home-page'); return false;"
     >
       <div class="d-inline-block">
         <%= inline_file("chevron-left.svg") %>
